fixed MQ partition
