+++ .gitignore
@@ -16,3 +16,4 @@ jtester.properties
 .DS_Store
 *.tar.gz
 *.rpm
+client-adapter/example/
