--- client/src/test/java/com/alibaba/otter/canal/client/running/rocketmq/AbstractRocektMQTest.java
@@ -1,9 +1,0 @@
-package com.alibaba.otter.canal.client.running.rocketmq;
-
-public class AbstractRocektMQTest {
-
-    public static String topic       = "example";
-    public static String groupId     = "group";
-    public static String nameServers = "localhost:9876";
-
-}
