--- example/src/main/java/com/alibaba/otter/canal/example/db/dialect/mysql/MysqlDialect.java
@@ -1,32 +1,0 @@
-/*
- * Copyright (C) 2010-2101 Alibaba Group Holding Limited.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.alibaba.otter.canal.example.db.dialect.mysql;
-
-import com.alibaba.otter.canal.example.db.dialect.AbstractDbDialect;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.support.lob.LobHandler;
-
-public class MysqlDialect extends AbstractDbDialect {
-
-    public MysqlDialect(JdbcTemplate jdbcTemplate, LobHandler lobHandler) {
-        super(jdbcTemplate, lobHandler);
-    }
-
-    public boolean isEmptyStringNulled() {
-        return false;
-    }
-}
