+++ pom.xml
@@ -294,7 +294,7 @@
             <dependency>
                 <groupId>mysql</groupId>
                 <artifactId>mysql-connector-java</artifactId>
-                <version>5.1.40</version>
+                <version>5.1.47</version>
                 <!--<scope>test</scope>-->
             </dependency>
             <dependency>
